requirements.txt
# === Core Backend ===
fastapi
uvicorn[standard]         
pydantic>=2.7.0            # latest supports Python 3.13
python-dotenv

# === Machine Learning ===
numpy>=2.1.0               # force NumPy 2.x
pandas>=2.2.2              # NumPy 2.x compatible
scikit-learn>=1.5.0        # official NumPy 2 support
xgboost>=2.1.0             # NumPy 2.x compatible
joblib>=1.4.0
openpyxl

# === Database ===
sqlalchemy>=2.0.30
alembic>=1.13.2
psycopg2-binary

# === IBM Watson / Granite ===
ibm-watson-machine-learning>=1.0.347
ibm-watsonx-ai>=0.6.4

# === Utils ===
# requests
# python-multipart
# cryptography

# === Testing ===
# pytest
# httpx
